# ewm-main-service/src/main/java/ru/practicum/dto/comment/UpdateCommentDto.java
package ru.practicum.dto.comment;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.hibernate.validator.constraints.Length;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UpdateCommentDto {
    @NotBlank
    @Length(min = 5, max = 2000)
    private String text;
}